Move scripts around - move into automation namespace
