+++ UnityPomodoro/Assembly-CSharp.csproj
@@ -71,7 +71,7 @@
      <Compile Include="Assets\AdrianMiasik\Scripts\Components\Specific\EndTimestampBubble.cs" />
      <Compile Include="Assets\AdrianMiasik\Scripts\Components\Core\Items\SidebarRow.cs" />
      <Compile Include="Assets\AdrianMiasik\Scripts\Components\Core\Items\DigitSeparator.cs" />
-     <Compile Include="Assets\AdrianMiasik\Scripts\Components\Specific\MediaCreator.cs" />
+     <Compile Include="Assets\AdrianMiasik\Scripts\Components\Specific\Automation\MediaCreator.cs" />
      <Compile Include="Assets\AdrianMiasik\Scripts\Components\Core\ClickButtonText.cs" />
      <Compile Include="Assets\AdrianMiasik\Scripts\Components\Base\Toggle.cs" />
      <Compile Include="Assets\AdrianMiasik\Scripts\Components\Core\ToggleSlider.cs" />
@@ -92,7 +92,7 @@
      <Compile Include="Assets\AdrianMiasik\Scripts\Components\Specific\DigitFormatDropdown.cs" />
      <Compile Include="Assets\AdrianMiasik\Scripts\Components\Specific\CreditsBubble.cs" />
      <Compile Include="Assets\AdrianMiasik\Scripts\Components\Specific\SkipButton.cs" />
-     <Compile Include="Assets\AdrianMiasik\Scripts\Components\Base\MediaCapture.cs" />
+     <Compile Include="Assets\AdrianMiasik\Scripts\Components\Specific\Automation\MediaCapture.cs" />
      <Compile Include="Assets\AdrianMiasik\Scripts\Interfaces\ITimerState.cs" />
      <Compile Include="Assets\AdrianMiasik\Scripts\Components\Core\Helpers\CollectionHelper.cs" />
      <None Include="Assets\TextMesh Pro\Shaders\TMPro.cginc" />
